# mok-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api"
    }
  ],
  "paths": {
...
    "/appointments/": {
      "get": {
        "tags": ["appointments"],
        "summary": "Get Appointments",
        "description": "Получить список записей на прием",
        "operationId": "get_appointments_appointments__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentBase"
                  },
                  "title": "Response Get Appointments Appointments  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/create": {
      "post": {
        "tags": ["appointments"],
        "summary": "Create Appointment",
        "description": "Создать новую запись на прием к психологу",
        "operationId": "create_appointment_appointments_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}": {
      "get": {
        "tags": ["appointments"],
        "summary": "Get Appointment",
        "description": "Получить информацию о конкретной записи",
        "operationId": "get_appointment_appointments__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}/cancel": {
      "put": {
        "tags": ["appointments"],
        "summary": "Cancel Appointment",
        "description": "Отменить запись на прием (для пациента)",
        "operationId": "cancel_appointment_appointments__id__cancel_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/roles/{user_id}/assign": {
      "post": {
        "tags": ["roles"],
        "summary": "Assign Role",
        "description": "Назначить роль пользователю",
        "operationId": "assign_role_roles__user_id__assign_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Assign Role Roles  User Id  Assign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roles/{user_id}/remove": {
      "post": {
        "tags": ["roles"],
        "summary": "Remove Role",
        "description": "Убрать роль у пользователя",
        "operationId": "remove_role_roles__user_id__remove_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRemoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Remove Role Roles  User Id  Remove Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/therapists/{psychologist_id}": {
      "get": {
        "tags": ["therapists"],
        "summary": "Get Psychologist",
        "description": "Получить информацию о конкретном психологе по ID",
        "operationId": "get_psychologist_therapists__psychologist_id__get",
        "parameters": [
          {
            "name": "psychologist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Psychologist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PsychologistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["therapists"],
        "summary": "Delete Psychologist Endpoint",
        "operationId": "delete_psychologist_endpoint_therapists__psychologist_id__delete",
        "parameters": [
          {
            "name": "psychologist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Psychologist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Psychologist Endpoint Therapists  Psychologist Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/therapists/": {
      "get": {
        "tags": ["therapists"],
        "summary": "Get Psychologists",
        "description": "Получить список всех психологов с пагинацией",
        "operationId": "get_psychologists_therapists__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Количество записей для пропуска",
              "default": 0,
              "title": "Skip"
            },
            "description": "Количество записей для пропуска"
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "exclusiveMinimum": 0,
              "description": "Количество записей для получения",
              "default": 10,
              "title": "Take"
            },
            "description": "Количество записей для получения"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PsychologistResponse"
                  },
                  "title": "Response Get Psychologists Therapists  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["therapists"],
        "summary": "Create Psychologist Endpoint",
        "operationId": "create_psychologist_endpoint_therapists__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PsychologistCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PsychologistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentBase": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "psychologist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Psychologist Id"
          },
          "type": {
            "$ref": "#/components/schemas/AppointmentType"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "scheduled_time": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Time"
          },
          "remind_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remind Time"
          },
          "last_change_time": {
            "type": "string",
            "format": "date-time",
            "title": "Last Change Time"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "psychologist_id",
          "type",
          "status",
          "scheduled_time",
          "last_change_time",
          "venue"
        ],
        "title": "AppointmentBase"
      },
      "AppointmentCancelRequest": {
        "properties": {
          "cancel_reason": {
            "type": "string",
            "title": "Cancel Reason"
          }
        },
        "type": "object",
        "required": ["cancel_reason"],
        "title": "AppointmentCancelRequest"
      },
      "AppointmentCreateRequest": {
        "properties": {
          "application_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "psychologist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Psychologist Id"
          },
          "type": {
            "$ref": "#/components/schemas/AppointmentType"
          },
          "scheduled_time": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Time"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "remind_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remind Time"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": ["patient_id", "psychologist_id", "type", "scheduled_time"],
        "title": "AppointmentCreateRequest"
      },
...
      "AppointmentStatus": {
        "type": "string",
        "enum": ["Approved", "Accepted", "Cancelled", "Done"],
        "title": "AppointmentStatus"
      },
      "AppointmentType": {
        "type": "string",
        "enum": ["Offline", "Online"],
        "title": "AppointmentType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PsychologistCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "experience": {
            "type": "string",
            "title": "Experience"
          },
          "qualification": {
            "type": "string",
            "title": "Qualification"
          },
          "consult_areas": {
            "type": "string",
            "title": "Consult Areas"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "office": {
            "type": "string",
            "title": "Office"
          },
          "education": {
            "type": "string",
            "title": "Education"
          },
          "short_description": {
            "type": "string",
            "title": "Short Description"
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "experience",
          "qualification",
          "consult_areas",
          "description",
          "office",
          "education",
          "short_description"
        ],
        "title": "PsychologistCreateRequest"
      },
      "PsychologistResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "experience": {
            "type": "string",
            "title": "Experience"
          },
          "qualification": {
            "type": "string",
            "title": "Qualification"
          },
          "consult_areas": {
            "type": "string",
            "title": "Consult Areas"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "office": {
            "type": "string",
            "title": "Office"
          },
          "education": {
            "type": "string",
            "title": "Education"
          },
          "short_description": {
            "type": "string",
            "title": "Short Description"
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "experience",
          "qualification",
          "consult_areas",
          "description",
          "office",
          "education",
          "short_description",
          "photo",
          "first_name",
          "middle_name",
          "last_name",
          "phone_number"
        ],
        "title": "PsychologistResponse"
      },
...
      "RoleAssignRequest": {
        "properties": {
          "role_code": {
            "$ref": "#/components/schemas/RoleCode"
          }
        },
        "type": "object",
        "required": ["role_code"],
        "title": "RoleAssignRequest"
      },
      "RoleCode": {
        "type": "string",
        "enum": ["user", "psychologist", "admin", "content_manager"],
        "title": "RoleCode"
      },
      "RoleRemoveRequest": {
        "properties": {
          "role_code": {
            "$ref": "#/components/schemas/RoleCode"
          }
        },
        "type": "object",
        "required": ["role_code"],
        "title": "RoleRemoveRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}